# src/app/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mohammadiya App API",
    "version": "1.0.0",
    "description": "API for Admin, Teachers, Students"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "ClassLevel": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "TimeSlot": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "studentName": { "type": "string" },
          "rollNumber": { "type": "string" },
          "grNumber": { "type": "string" },
          "gender": { "type": "string" },
          "age": { "type": "integer" },
          "fatherName": { "type": "string" },
          "mobile": { "type": "string" },
          "formNo": { "type": "string" },
          "admissionFee": { "type": "number" },
          "monthlyFees": { "type": "number" },
          "remarks": { "type": "string" },
          "hafizCategory": { "type": "string" },
          "fullTimeSubCategory": { "type": "string" },
          "classSession": { "$ref": "#/components/schemas/ClassSession" }
        }
      },
      "ClassSession": {
        "type": "object",
        "properties": {
             "id": { "type": "string" },
             "classLevel": { "$ref": "#/components/schemas/ClassLevel" },
             "timeSlot": { "$ref": "#/components/schemas/TimeSlot" }
        }
      },
...
    }
  },
  "paths": {
    "/students": {
      "get": {
        "summary": "List Students with Filters",
        "tags": ["Students"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "classId", "in": "query", "schema": { "type": "string" } },
          { "name": "timeSlotId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of Students",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Student" } } } }
          }
        }
      },
      "post": {
        "summary": "Admission (Create Student)",
        "tags": ["Students"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["studentName", "fatherName", "mobile", "dateOfBirth", "gender"],
                "properties": {
                  "studentName": { "type": "string" },
                  "fatherName": { "type": "string" },
                  "gender": { "type": "string" },
                  "mobile": { "type": "string" },
                  "dateOfBirth": { "type": "string", "format": "date-time" },
                  "grNumber": { "type": "string" },
                  "rollNumber": { "type": "string" },
                  "residence": { "type": "string" },
                  "fullPermanentAddress": { "type": "string" },
                  "emergencyContactName": { "type": "string" },
                  "emergencyContactPhone": { "type": "string" },
                  "classId": { "type": "string" },
                  "timeSlotId": { "type": "string" },
                  "hafizCategory": { "type": "string" },
                  "fullTimeSubCategory": { "type": "string" },
                  "formNo": { "type": "string" },
                  "admissionFee": { "type": "number" },
                  "monthlyFees": { "type": "number" },
                  "remarks": { "type": "string" },
                  "age": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Student Created" }
        }
      }
    },
...
    "/students/promote": {
      "put": {
        "summary": "Bulk Promote Students",
        "tags": ["Students"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["studentIds", "classId", "timeSlotId"],
                "properties": {
                  "studentIds": { "type": "array", "items": { "type": "string" } },
                  "classId": { "type": "string" },
                  "timeSlotId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Promotion successful" }
        }
      }
    },
...
    "/students/{id}": {
      "get": {
        "summary": "Get Student Details",
        "tags": ["Students"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Student details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Student" } } } }
        }
      },
      "put": {
        "summary": "Update Student Details",
        "tags": ["Students"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "studentName": { "type": "string" }, "status": { "type": "string" } } } 
            }
          }
        },
        "responses": {
          "200": { "description": "Student updated" }
        }
      }
    },
...
  }
}